+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 =====
 
 *   (bc) Adapt to new `_dependencies.json` storage directory + new layout.
+*   (bc) CSS files are not hashed anymore, as they are already hashed in Firefly.
 
 
 1.0.1